% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioDataScience3-package.R
\docType{package}
\name{BioDataScience3-package}
\alias{BioDataScience3-package}
\title{A series of learnr tutorials and Shiny applications for BioDataScience3}
\description{
Run the tutorials using\code{\link[=run]{run()}}.
}
