% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr.R
\name{learnr_setup}
\alias{learnr_setup}
\alias{learnr_banner}
\alias{learnr_server}
\title{Prepare a learnr tutorial for learndown recording}
\usage{
learnr_setup(cap = "Code R", debug = Sys.getenv("LEARNDOWN_DEBUG", 0) != 0)

learnr_banner(simple = FALSE)

learnr_server(input, output, session)
}
\arguments{
\item{cap}{The caption of learnr R code widgets.}

\item{debug}{Do we issue debugging messages (by default, yes if the
environment variables \code{LEARNDOWN_DEBUG} is not \code{0}).}

\item{simple}{Do we use a simple banner without title or not?}

\item{input}{The Shiny input.}

\item{output}{The Shiny output.}

\item{session}{The Shiny session.}
}
\description{
These functions add recording functionalities through the learndown approach.
}
